move subPackPipe to engine and cache asset slowly (#72)
* move subPackPipe to engine

* use asset directly, then cache it slowly

refine

+++ wechatgame/libs/wx-downloader.js
@@ -35,6 +35,12 @@ var binary_format = [
 
 const REGEX = /^\w+:\/\/.*/;
 
+// used to control cache
+var cacheQueue = {};
+var checkNextPeriod = false;
+// cache one per cycle
+var cachePeriod = 100;
+
 var fs = wx.getFileSystemManager ? wx.getFileSystemManager() : null;
 
 var _newAssets = [];
@@ -185,12 +191,17 @@ function nextPipe(item, callback) {
     var queue = cc.LoadingItems.getQueue(item);
     queue.addListener(item.id, function (item) {
         if (item.error) {
-            fs && fs.unlink({
-                filePath: item.url,
-                success: function () {
-                    cc.log('Load failed, removed local file ' + item.url + ' successfully!');
-                }
-            });
+            if (item.url in cacheQueue) {
+                delete cacheQueue[item.url];
+            }
+            else {
+                fs && fs.unlink({
+                    filePath: item.url,
+                    success: function () {
+                        cc.log('Load failed, removed local file ' + item.url + ' successfully!');
+                    }
+                });
+            }
         }
     });
     callback(null, null);
@@ -322,6 +333,38 @@ function ensureDirFor (path, callback) {
     });
 }
 
+function cacheAsset (url, localPath) {
+    cacheQueue[url] = localPath;
+
+    if (!checkNextPeriod) {
+        checkNextPeriod = true;
+        function cache () {
+            checkNextPeriod = false;
+            for (var url in cacheQueue) {
+                var localPath = cacheQueue[url];
+                ensureDirFor(localPath, function () {
+                    // Save to local path
+                    wx.saveFile({
+                        tempFilePath: url,
+                        filePath: localPath,
+                        success: function (res) {
+                            cc.log('cache success ' + localPath);
+                        }
+                    });
+                });
+                
+                delete cacheQueue[url];
+                if (!cc.js.isEmptyObject(cacheQueue) && !checkNextPeriod) {
+                    checkNextPeriod = true;
+                    setTimeout(cache, cachePeriod);
+                }
+                return;
+            }
+        };
+        setTimeout(cache, cachePeriod);
+    }
+}
+
 function downloadRemoteFile (item, callback) {
     // Download from remote server
     var relatUrl = item.url;
@@ -340,37 +383,15 @@ function downloadRemoteFile (item, callback) {
             if (res.statusCode === 200 && res.tempFilePath) {
                 // http reading is not cached
                 var temp = res.tempFilePath;
-                var localPath = wx.env.USER_DATA_PATH + '/' + relatUrl;
-                // check and mkdir remote folder has exists
-                ensureDirFor(localPath, function () {
-                    // Save to local path
-                    wx.saveFile({
-                        tempFilePath: res.tempFilePath,
-                        filePath: localPath,
-                        success: function (res) {
-                            // cc.log('save:' + localPath);
-                            item.url = res.savedFilePath;
-                            if (item.type && non_text_format.indexOf(item.type) !== -1) {
-                                nextPipe(item, callback);
-                            }
-                            else {
-                                readText(item, callback);
-                            }
-                        },
-                        fail: function (res) {
-                            // Failed to save file, then just use temp
-                            console.log(res && res.errMsg ? res.errMsg : 'save file failed: ' + remoteUrl);
-                            console.log('It might be due to out of storage spaces, you can clean your storage spaces manually.');
-                            item.url = temp;
-                            if (item.type && non_text_format.indexOf(item.type) !== -1) {
-                                nextPipe(item, callback);
-                            }
-                            else {
-                                readText(item, callback);
-                            }
-                        }
-                    });
-                });
+                item.url = temp;
+                if (item.type && non_text_format.indexOf(item.type) !== -1) {
+                    nextPipe(item, callback);
+                }
+                else {
+                    readText(item, callback);
+                }
+                cacheAsset(temp, wx.env.USER_DATA_PATH + '/' + relatUrl);
+                
             }
             else {
                 cc.warn("Download file failed: " + remoteUrl);
@@ -387,7 +408,7 @@ function downloadRemoteFile (item, callback) {
                 errorMessage: res && res.errMsg ? res.errMsg : "Download file failed: " + remoteUrl
             }, null);
         }
-    })
+    });
 }
 
 // function downloadRemoteTextFile (item, callback) {
